Moved run_system() function
