+++ src/fastoad/models/performances/mission/openmdao/tests/test_payload_range.py
@@ -17,8 +17,8 @@ import numpy as np
 import pytest
 from numpy.testing import assert_allclose
 
-from fastoad._utils.testing import run_system
 from fastoad.io import DataFile
+from fastoad.testing import run_system
 from ..payload_range import PayloadRange
 
 DATA_FOLDER_PATH = Path(__file__).parent / "data"
