+++ src/fastoad/models/performances/mission/openmdao/tests/test_sizing_breguet.py
@@ -18,7 +18,7 @@ import openmdao.api as om
 import pytest
 from numpy.testing import assert_allclose
 
-from fastoad._utils.testing import run_system
+from fastoad.testing import run_system
 from ..mission import OMMission
 
 RESULTS_FOLDER_PATH = Path(__file__).parent / "results" / Path(__file__).stem
